Implemented the modules Todo_list and Preprocessor
Now we handle the 'pdfinc' and 'cite' properly.
As we do not accept closing braces ('}'), there should not be too much
sanitization trouble, but it should be checked.
There is an implementation of nsplit taken from batteries, I still don't
understand why it works that way, and loops infinitely when in
batteries.