Improve Mix.Local_bibliography
This modification even though a bit hacky, allows to choose your own
identifier for the bibliographic references. The `cite' function can
be called with an ?id optional argument to register use and it.  To be
then able to register an id without actually printing anything there
is the ?silent argument (hence then the id is registered when using
the `cites' function).