Collapse directories and pages by name in menus
Now if there is a page Page.brtx in a directory called Page, the link
will be on the `directory' part of the menu.
The patch also adds the dibrawimenudirname CSS class to add some
visibility to directory names even when they are links.
There is a default style for it in the HTML_template module:
  .dibrawimenudirname { font-size: 115%;  font-weight:bold; }